Initial stab at beginnings of composition
My goal with this was to start figuring out the abstractions in such a
way that it would be clear and easy to implement things in pure
isolation but also be able to compose them together to make more
complicated larger things.

This will allow developers to be able to develop view components or
flows in complete isolation and test it in all the various states
without any issues and then after integrate it into the app.