Add ability for a component to have multiple stores
Add ability for a component to have multiple stores so that it could for
example get some of it state as borrowed state from a parent component
and also have other state that it fully owns.

As this does work, it does feel like this is now putting the
responsibility of somewhat of an integration concern into the component
as it has to know that some state is coming from a parent and other
state is not.

So this likely needs to be revisited but this is at least a start even
if it takes use to slightly less pure place.