Add ComponentWidget & ComponentStatefulWidget
Add ComponentWidget & ComponentStatefulWidget so that we have a simpler
interfaces that are more along the lines of what Flutter devs would
expect.

I also updated all the examples appropriately to use the widget rather
than the probably soon to be depricated viewBuilder() interface.

I also changed the typing through flutter_forge so that instead of
`State` I am now using `S`. This is because I ran into name collisions
with `State` and it was causing confusion.

ps-id: d39892e6-7587-447b-8f5e-b81a71185952