Split action and side effect appart.
Split action and side effect appart so that state changes are 100% pure
and so that the side effecting code is pushed up toward the outer
boundary of the application. This will in turn make things more
testable and make it so that users have less opportunity to make
mistakes in their actions.

ps-id: 060dddaf-459d-4bf6-9399-67a49567a30b