Add LoadOnInitComponent scenario
Add LoadOnInitComponent scenario so that we could test the scenario of
having a component load state on init. Right now this scenario is
causing an exception when it is loaded.

The goal of this scenario though is to help us formalize state
initialization within a component launch in a way where users can't run
into these exceptions.

ps-id: 3beed196-76a3-4ae3-97fd-4947e7a7c999