Add postInitState() lifecycle to ComponentWidget
Add postInitState() lifecycle to ComponentWidget so that we could safely
use it as a hook to send actions to the store when the widget's
associated state is initialized the first time.

I left the initState() lifecycle in the ComponentWidget as I didn't want
to take away classic flutter hooks that people are already used to and I
didn't want to change the behavior of classic flutter hooks that people
are already used to.

This basically makes it so that the initState() hook is called and then
immediately following that the postInitState() hook is called in a way
where it doesn't get executed until after the widget is actually built
for the first time.

[changelog]
added: postInitState() hook to the ComponentWidget

ps-id: 9a3b44ed-b9ac-4310-8ba4-8f8a7dc69618