Remove the send action to child store example
Remove the send action to child store example because it isn't a valid
example. If a parent component wants to impact the state of a child
component then the parent component needs to have that state as part of
it's State class and scope a store down to the child component to use.
Then it can impact the state it owns via it's own actions and the child
component's scoped store can trigger updates when appropriate.

ps-id: e92d8d3e-6a48-46e2-b020-e4fc38ee9530