Remove viewStore param from ComponentWidget dispose() hook
Remove viewStore param from ComponentWidget dispose() hook because it
was causing an exception. Specifically it was causing the following
exception.

══╡ EXCEPTION CAUGHT BY WIDGETS LIBRARY ╞═══════════════════════════════════════════════════════════
The following assertion was thrown while finalizing the widget tree:
Looking up a deactivated widget's ancestor is unsafe.
At this point the state of the widget's element tree is no longer stable.
To safely refer to a widget's ancestor in its dispose() method, save a reference to the ancestor by
calling dependOnInheritedWidgetOfExactType() in the widget's didChangeDependencies() method.

When the exception was thrown, this was the stack:
0      Element._debugCheckStateIsActiveForAncestorLookup.<anonymous closure> (package:flutter/src/widgets/framework.dart:4241:9)
1      Element._debugCheckStateIsActiveForAncestorLookup (package:flutter/src/widgets/framework.dart:4255:6)
2      Element.getElementForInheritedWidgetOfExactType (package:flutter/src/widgets/framework.dart:4286:12)
3      ProviderScope.containerOf (package:flutter_riverpod/src/framework.dart:99:12)
4      ConsumerStatefulElement.read (package:flutter_riverpod/src/consumer.dart:552:26)
5      Store.viewStore (package:flutter_forge/src/store.dart:23:16)
6      _ComponentState.dispose (package:flutter_forge/src/component_widget.dart:46:26)
7      StatefulElement.unmount (package:flutter/src/widgets/framework.dart:5105:11)

[changelog]
changed: ComponentWidget dispose() hook to have no params

ps-id: abf3cd4d-8478-4cb7-b7c2-251201a0e4db