Move from selfContained to default value constructor
Up to this point in our examples we were using a factory constructor to
with the naming convention of selfContained to create an instance of the
widget that owned it's own store. This was because we tried to use
default value constructor but it only works if the value is a constant.
A store is not a constant and therefore we couldn't use that.

However, we did discover that we can just call super() and implement our
own default value construction. We believe this is a better approach
than having a naming convention with a factory consntructor. Therefore
we have shifted all of our examples to now use this approach for
creating a self contained widget.

ps-id: 7d8a2532-747d-4e37-9d10-6e96944b7d4b