Add ReducerTuple.noop factory constructor
Add ReducerTuple.noop factory constructor so that we could easily and
meaningfully create noop reducer/action tuples. This is extremely useful
in the scenario where you want to make a component that defines actions
but doesn't want them to change any state.

This is often scenarios where you say have a button inside of a
component and you want the button tapped action to exist and be called
by the component but you want whatever the parent is to be able to
handle the state change and effect of that action.

[changelog]
added: ReducerTuple/ActionTuple.noop() factory constructor

ps-id: d60d0235-522b-4658-a20d-2947444993b5