Change EffectTask from typealias to a class
Change EffectTask from typealias to a class because I thought it makes
things a little bit clearer and more explicit when creating EffectTasks
rather than just creating an anonymous function. Additionaly it enables
us to move the pullback() functionality to the class which makes a lot
of sense conceptually and when also trying to deal with the optionality
of them in dart.

This has enabled us to clean things up a bit and remove the
mapEffectTask() function as we no longer need it thanks to the
ergonomics of being able to do things like the following.

effectTask?.pullback()

[changelog]
changed: EffectTask from typealias of func to a class
changed: pullbackEffectTask() to EffectTask.pullback()
removed: mapEffectTask()

ps-id: e66efeed-e17a-4ca8-af3b-e9775bd0feef