Make ReducerTuple have a list of EffectTasks
Make ReducerTuple have a list of EffectTasks instead of just having one
optional EffectTask. This gives us more flexibility when defining our
actions to have them return a list of EffectTasks.

More importantly though it gives us the ability to compose actions
together in the future. It does so by eliminating a dependency between
ReducerActions and EffectTasks in terms of their composition operations.

For example prior to this change if we attempted to implement compose of
ReducerActions it would required us to be able to compose EffectTasks
and EffectTasks compose function requires the ability to compose
ReducerActions. This looping nature seemed bad and switching to a
collection of EffectTasks breaks that loop and gives us more
flexibility.

[changelog]
changed: ReducerTuple.effectTask to ReducerTuple.effectTasks

ps-id: f1cc26ed-457c-4b5c-bed1-74cedb5361f0