Add BuildContext to EffectTask parameters
Add BuildContext to EffectTask parameters so that effects could do
things like Navigator.of(context) in the future without having to do
gymnastics like writing function builders which people seem to have a
hard time with for some reason initially.

To accomplish this I had to use the setContext() function in the
ComponentWidget to initially set the BuildContext in the view store that
it can be passed in as an argument when executing the effects. I of
course also then had to pass it in where the effect is executed. Beyond
that I had to modify the EffectTask.pullback to account for this
additional parameter as well.

[changelog]
changed: EffectTask to include BuildContext as a parameter

ps-id: 134a4efa-0f59-4425-8912-651dda31a81e