Convert Actions over to be data object based
I did this because we couldn't properly uniquely identify the actions
when they were represented as functions and they had different
paramaters. Switching back to a more classic messaging style data model
approach gives us that required flexibility. Normally we would want to
use union types or at least enums with associated values but Dart
doesn't really support that stuff to the extent that we would need it.
So until then we are opting to use this class based approach.

I also removed the viewBuilder to simplify things and focus on the
ComponentWidget which is really how we want to people to be use the
framework.

In addition I move all the core members of the framework into the
lib/src/core so that we could easily manage the higher level stuff above
that and create a little bit more structure.

These changes are breaking changes and also impact some of the necessary
typing requirements.

[changelog]
changed: Actions to be class based
removed: viewBuilder functionality

ps-id: 4651ea52-94f3-44c5-94c2-3dce01e55f93