Organize State Management code
Organize State Management code so that people coming into the code base
can easily understand what that the purpose of all of that code is to
facilitate state management in a controlled way on top of Riverpod. And
also to understand that if someone just wanted to use that piece of the
puzzle without the other benefits because they just wanted controlled
statement management with redux on top of Riverpod, they can.