WIP: add ability to bridge Riverpod providers
Add the ability to bridge a Riverpod provider into Flutter Forge so that
we could use Riverpod to fetch and cache data more easily.

One thing that needs to be thought through is the modularization &
isolation of the component when using external Riverpod providers this
way.

We may decide that this mechanism isn't what we want and we want a
completely different startegy.

This is just a first swag at how this might be possible.

<!-- ps-id: fdc912c2-a806-4cf0-9a6f-1077f863c55b -->