Remove Riverpod dependency from flutter_forge
Remove Riverpod dependency from flutter_forge so that we wouldn't have
any issues with version conflicts between applications using Riverpod +
flutter_forge together.

The way I accomplished this was essentially to by changing the ViewStore
concept to be a ChangeNotifier and using Flutter's ChangeNotifier
concept to facilitate implementing state change observations.

Making this work required also reimplementing the `scope` operation on
the Store as well as including the environment conversion in the `scope`
operation which in term adds the generic typeing to the store interface.

<!-- ps-id: 178373d2-e9e0-48bf-8d49-8db991dbbfc8 -->