Make BuildContext on a ViewStore optional
Make BuildContext on a ViewStore optional so that we can effectively
have Stores that aren't associated with a particular ComponentWidget.

To accomplish this I simply made a context property on ViewStore that is
typed as BuildContext?. I then removed the functions that were part of
the ViewStoreInterface for setting and getting the previously `late`
initialized _context property on the ViewStore. Then I updated the
places that were necessary to make sure that it was optional and the
optionality was dealt with all the way up to the EffectTask.

This enables us to be able to stores indepedent of ComponentWidgets that
can fundamentally stand and operate on their own. The side effect of
that is that the effects that are run by that store won't have access to
a BuildContext because the store isn't bound to a ComponentWidget.

<!-- ps-id: 8445c1eb-8658-429d-9be6-d0382b18d378 -->