Fix memory leak & prevention of finalizer runs
Fix memory leak & prevention of finalizer running. Specifically the
finalizer to unlisten was never getting executed because it is wired to
be triggered based on the childStore being freed. However, the
childStore was never getting free'd because the syncParentToChildState
lambda was capturing a strong reference to the childStore.

Switching the reference to a weak reference in that lambda allows the
childStore to properly be free'd which in turn now causes the finalizer
that unlistens from the parent viewstore to facilitate the syncing to
the child store.

Gotta watch out for implicit strong captures when dealing with lambdas.

<!-- ps-id: 2426a184-30cf-4275-b638-645412282eef -->