Add constraint that State has to be Equatable
Add constraint that State has to be Equatable so that the typing will
cause the user to do the required thing and make the state classes
Equatable.

My approach to this was simply adding the `extends Equatable` constraint
to the various generic State types within the library. Then I updated
all of the examples to convert the state classes to be consistently
equatable.

<!-- ps-id: bbaca02f-8a07-4f70-b4c8-3ceafd580d05 -->