Simplify how components are composed
Up to this point in time components were composed by doing two things.

1. Scoping a Store down to a Store that is only concerned with what the
   child Component needs to worry about.
2. Pulling back the child Reducer up to the level of the Parent
   Component and combining it with the existing parent Reducer.

This composition was done this way because we followed existing
frameworks that do it this way. This is fundamentally due to the how the
`scope()` operation is implemented. It is impleneted in such a way that
it builds a new Reducer that effectively converts child Actions into
parent Actions and sends them up to the parent Store to be handled and
then synchronizes state changes from the parent back down to the childs
copy of state.

This approach functionally works fine. However, it does require that the
parent handle all of the actions of the child. Replicating all that
functionality is crazy which is where the process of pulling back a
child reducer up to the parent level came from.

But what if we approach it a different way. Where the child Store
actually managed its own Actions and its own State directly and then we
simply synchronized the state change up to the parent and vice versa. If
we could do that then we wouldn't have to worry about pulling back
reducers and combining reducers at all. In fact we also wouldn't have to
worry about providing a conversion between child Action and parent
Action either. It would simplify the act of composing components
significantly.

What about the cases where we do want the parent to receive the child
Action so that the parent can do additional work as result. Well we can
still take this approach but simply internally create an intermediate
Reducer that effectively acts as a proxy but also translates the child
Action to the parent Action and forwards it up to the parent store. This
can simply be a specialized scoping operation.

This is exactly what this patch does. I accomplished this by doing the
following.

- renamed Store `scope()` to `scopeParentHandles()`
- added Store `scopeSyncState()`
- added Store `scopeForwardActionsAndSyncState()`
- added `debug()` to Reducer so we could easily debug state mutation
- updated the examples to use the new scope functions as well as to
  print when widgets are built and disposed as well as to use the new
  reducer debug() so that we can see as the reducers receive and compute
  new state
- made both parent and child store references weak in the sync lambas

<!-- ps-id: efa2ee5a-2781-4ea8-9022-b5f3e829fd46 -->