Add AsyncState classes
Add AsyncState classes so that we could represent asynchronous state
within Flutter Forge and be able to easily and exhaustively evaluate it
within our widgets.

This allows us to represent state that is loaded asynchronously in a
consistent fashion.

final AsyncState<Car> car;

Then `car` can be in one of the following values.

- `AsyncState.initial()`
- `AsyncState.loading()`
- `AsyncState.data(Car())`
- `AsyncState.error(Object, StackTrace)`

Then we can exhaustively handle the cases of it as follows.

final widget = car.when(
  initial: () => const Text('Initial'),
  loading: () => const Text('Loading'),
  data: (v) => Text('Date = $v'),
  errer: (e, stack) => Text('Error = $e, $stack')
)

To accomplish this I created an abstract base class `AsyncState` which
provides factory functions to construct the four distinct subclasses
that represent each of the states, `AsyncStateInitial`,
`AsyncStateLoading`, `AsyncStateData`, and `AsyncStateError`.
`AsyncState` also defines property getters that the subclasses are
required to implement to make conditional checking quicker than always
doing a type check. In addition the `AsyncState` class provides a
`when()` that facilitates exhaustive case handling for an `AsyncState`.

[changelog]
added: AsyncState and related subclasses

<!-- ps-id: 03654b18-fd70-4159-af46-ddf14bc076a1 -->