Add AsyncStateWidget
Add AsyncStateWidget so that it could be used in combination with store
scoping in the future to facilitate having a nice way to tie an
asynchronous loading of a value without having to define all the actions
yourself and the reducer yourself.

To accomplish this I created a few things.

First I standardized a set of Actions for managing AsyncState.
Specifically, the AsyncStateAction abstract base class and its
associated subclasses, AsyncStateLoadAction,
AsyncStateLoadCompleteAction, AsyncStateLoadFailedAction,
AsyncStateLoadResetAction.

Secondly, I implemented the asyncStateReducer builder that builds a
reducer that properly manages state changes for an AsyncState value
based on these standardized Actions.

Thirdly, I created the AsyncStateWidget class which is a Flutter Forge
ComponentWidget that takes up no visible space. It is used solely for
the `postInitState` method to send the `AsyncStateLoadAction` to the
store and to facilitate easy store & state binding through scoping.

[changelog]
added: AsyncStateWidget to facilitate automated AsyncState loading

<!-- ps-id: 22e5d946-1ec1-4fe5-b151-86d39bf24a03 -->