Add loadAsyncStateWidget() to StoreInterface
Add loadAsyncStateWidget() to StoreInterface so that users would
be able to easily create a AsyncStateWidget from a store without having
to worry about setting up a reducer or setting up the environment
pass through. This is a convenience feature for sure but I think it
definitely makes a difference.

[changelog]
added: loadAsyncStateWidget() to all stores

<!-- ps-id: 25284e18-68a4-4cc7-97b1-14aa57fa6573 -->