Replace loadAsyncStateWidget w/ scopeAsyncStateSync
Replace loadAsyncStateWidget w/ scopeAsyncStateSync so that we don't
have the Store returning a Widget from a function which could have perf
impacts and also seems weird and inconsistent with the developer
experience. Instead it makes more sense to have just a specific scoping
method on the store that makes it easier to scope for AsyncStateWidget.

This gives us the best of both worlds as it removes the perf issue
possibility as well as maintains the developer experience of how a user
would expect to work with widgets.

<!-- ps-id: a606b70f-f1c4-4fe3-960a-194f597f101a -->