Modify load on component init to use environment
Modify load on component init to use environment so that we at least
have a basic example of using the environment to formalize dependency
injection. This will also set us up for a test in the future where we
can stub out this dependency injection using the environment in our
tests.

<!-- ps-id: 4690015f-cf45-454b-9811-a338373821c6 -->