Add TestStore
Add TestStore so that it could be used in tests to easily verify the
logic of not only the reducer but also the effects tied to it. So
effectively testing the behavior of the store and how it receives
actions and how those actions change state.

[changelog]
added: TestStore class to facilitate testing reducer & effect logic

<!-- ps-id: c04c9ce7-65b3-41da-9888-540233345134 -->