Disable ComponentWidget rebuild on state change
Disable ComponentWidget rebuild on state change because I have
determined that it as a default behavior is dangerous from a performance
standpoint especially when doing composition of components. Therefore,
I have changed the ComponentWidget to no longer automatically build on
state change and instead have the developer use the newly added
Rebuilder widget to facilitate rebuilding a very targeted portion of
the widget tree rather than everything in.

I also started adding some documentation.

My approach to doing this was to first remove the ValueListenableBuilder
usage in the build function of ComponentWidget. This eliminated the
rebuilding when the state changes. I then also removed the `state`
argument from the `build` method as having it there makes it seem like
it is going to update based on state change. Then I added the
`Rebuilder` widget so that developers could use that to watch state and
rebuild a portion of the widget tree when state changes. Then I simply
updated the examples to abide by these changes.

skip_ticket_id_check

<!-- ps-id: 0becabb9-ed01-41c5-a98b-f454f855a764 -->