Add ability to register listener on viewStore
Add ability to register listener on viewStore so that when the state
changes we can trigger arbitrary code. This is useful for things like
displaying a snackbar or an alert, etc.

To accomplish this I simply added a listen() method to the ViewStore and
ViewStoreInterface in which you pass a function that receives state and
is executed when the state has changed.

Then I updated the counter example to include a listen.

skip_ticket_id_check

<!-- ps-id: 8e4db505-3677-4ea4-8f45-0d0527385f61 -->