Change viewStore.listen to widget lifecycle method
Change viewStore.listen to widget lifecycle method because the use case
we are trying to support is that of impacting a widget programmatically
when state changes. Doing things like showing a snackbar, etc.
Therefore, the thinking is that having the listen bound to the widget
lifecycle allows us to cleanup the listener for the user as well as
listen without them having to worry about anything. This effectively
makes it so that they can't messup the cleanup of the listeners.

skip_ticket_id_check

<!-- ps-id: f83d8deb-fa08-46f7-879e-34f4cbd034bc -->