Add linting
Add linting so that we get better guidance in terms of writing our code
consistently.

The approach I took was to add the analysis_options.yaml to the root and
to add the dependencies to the pubspec.yaml and then to fix or
explicitly ignore the cases where we have to go against the linter.

<!-- ps-id: 8a931a8c-62a6-4e11-858f-9291a6fadcd6 -->