Add SelectRebuilder widget
Add SelectRebuilder widget so that users would be able to explicitly
trigger a rebuild of a portion of their widget tree when a particular
selection of state changes rather than when any of a components state
changes.

This is primarily enables optimization as all child components should be
able to properly be able to be recreated from state. I modified the
load_on_component_init example to use the SelectRebuilder widget rather
than a single Rebuilder widget.

The approach I took to implement this was first to implement a
SelectedValueNotifier to facilitate observing and being notified based
on a selection of state. Then I used that within the implementation of
SelectRebuilder to be able to trigger build based on the intended
selection.

skip_ticket_id_check

[changelog]
added: SelectRebuilder widget

<!-- ps-id: 1d661340-ca61-4797-a9e0-69a6dc57257f -->