Organize the code base
Organize the code base so that it is split into the following sections.

- lib/src/convenience - higher level convenience features
- lib/src/core/state_management - the core state management, Store, etc.
- lib/src/core/widgets - widgets to facilate moduler & compossible state
  based widget building using state_management
- lib/src/utils - generic utilities we have created and use throughout

The goal with this was simply to organize the code base in a better
fashion so that other people would be able to get a general
understanding of the major pieces of the libary and know where things
should belong.

<!-- ps-id: 1d37321d-0a76-4db3-92b9-5bac1de71149 -->