Improve Riverpod integration
Improve Riverpod integration so that the confusing part of the
integration is encapsulated behind a Riverpod widget. This should make
it easier for people to bridge between the worlds as well as it should
prevent users from inadvertently causing more rebuilds than needed.

This Widget probably should live somewhere other than examples. But we
decided that we do not want the Flutter Forge library to have any
dependence on Riverpod because of concern of lock in and maintenance as
Riverpod evolves. So maybe this is just something we provide in a Guide
and can be copy and pasted into peoples projects if they need it.

<!-- ps-id: 8ff820af-89b8-4a57-ac59-373667bcf893 -->