Add optional builder to ComponentWidget
Add optional builder argument to ComponentWidget constructor so that
users can override the UI of a component via the optional builder
argument if they prefer that approach over creating a widget that
inherits from the other widget and then overrides the body method.

This involved add a the optional `builder` function to the
ComponentWidget constructor and then adding logic to the
ComponentWidget's `build` method to check if `builder` is provided, if
so use it for the `body`, otherwise fall back to the widgets provided
`body`.

After doing this I was running into run time type checking failures and
I realized that the typing wasn't as explicit as it could be in
ComponentWidget. Therefore, I made the typing more explicit in the
definition of _ComponentState so that the relationship between
_ComponentState generic types and ComponentWidget's generic types are
properly bound. This solved the issue and should also solve issues where
in some scenarios inference was previously resulting in dynamic for
types.

Beyond those changes I added an example to facilitate seeing this in and
app and I also added a test to cover this feature of the ComponentWidget
as well. As part of the example work I created a separate counter
component that doesn't own it's own store and one that does to help
create more of a distinction between those cases. Then I updated the
respective examples to use the appropriate counter components.

I also updated the documentation of the ComponentWidget to describe this
feature.

This addresses issue #18.

[changelog]
added: optional `builder` arg to ComponentWidget for overriding ui
fixed: typing of ComponentWidget to better provide type inference

<!-- ps-id: 4862765c-9d8e-4def-a85f-350b434a8426 -->