Fix bug in AsyncState.when() error case
Fix bug in AsyncState.when() error case as it was incorrectly passing
the error closure in as the error object. This was due to a name
collision between the `error` getter on the object and the `error`
argument in the `when` method.

To address this I explicitly reference the error object stored within
the AsyncStateError type using the `this.error` accessor rather than the
short form of `error`.

skip_ticket_id_check

[changelog]
fixed: AsyncState.when(errror:) passing incorrect error object

<!-- ps-id: 4adf8702-d70a-4035-bc05-5f88a1f8a36b -->